e2e-demos/savepoint-demo/requirements.txt
confluent-kafka
pydantic
pyyaml
pydantic-avro
fastavro